+++ R/aaa.R
@@ -69,6 +69,10 @@ the <- new.env(parent = emptyenv())
 #' `on_load()` is implemented via `.onLoad()` and requires
 #' `run_on_load()` to be called from that hook.
 #'
+#' The expressions inside `on_load()` do not undergo static analysis
+#' by `R CMD check`. Therefore, it is advisable to only use
+#' simple function calls inside `on_load()`.
+#'
 #' @examples
 #' quote({  # Not run
 #' 
@@ -79,16 +83,20 @@ the <- new.env(parent = emptyenv())
 #' }
 #'
 #' # Register a method on load
-#' on_load(s3_register("foo::bar", "my_class"))
+#' on_load({
+#'   s3_register("foo::bar", "my_class")
+#' })
 #'
 #' # Assign an object on load
 #' var <- NULL
-#' on_load(
+#' on_load({
 #'   var <- foo()
-#' )
+#' })
 #'
 #' # To use `on_package_load()` at top level, wrap it in `on_load()`
-#' on_load(on_package_load("foo", message("foo is loaded")))
+#' on_load({
+#'   on_package_load("foo", message("foo is loaded"))
+#' })
 #'
 #' # In functions it can be called directly
 #' f <- function() on_package_load("foo", message("foo is loaded"))