+++ src/rlang/altrep.h
@@ -1,6 +1,9 @@
+// IWYU pragma: private; include "rlang.h"
+
 #ifndef RLANG_ALTREP_H
 #define RLANG_ALTREP_H
 
+#include "rlang-types.h"
 
 #if (R_VERSION < R_Version(3, 5, 0)) ||                 \
   (defined(_WIN32) && R_VERSION == R_Version(3, 5, 0))