+++ src/rlang/c-utils.h
@@ -1,6 +1,11 @@
+// IWYU pragma: private; include "rlang.h"
+
 #ifndef RLANG_C_UTILS_H
 #define RLANG_C_UTILS_H
 
+#include "rlang-types.h"
+
+#include <inttypes.h>
 #include <math.h>
 #include <float.h>
 #include "cnd.h"