+++ src/rlang/dyn-array.h
@@ -1,6 +1,11 @@
+// IWYU pragma: private; include "rlang.h"
+
 #ifndef RLANG_DYN_ARRAY_H
 #define RLANG_DYN_ARRAY_H
 
+#include "rlang-types.h"
+#include "c-utils.h"
+#include "cnd.h"
 #include "vec.h"
 
 
@@ -172,6 +177,18 @@ static inline
 void r_dyn_cpl_push_back(struct r_dyn_array* p_vec, r_complex elt) {
   r_ssize loc = r__dyn_increment(p_vec);
   r_dyn_cpl_poke(p_vec, loc, elt);
+}
+static inline
+void r_dyn_raw_push_back(struct r_dyn_array* p_vec, char elt) {
+  r_ssize loc = r__dyn_increment(p_vec);
+  r_dyn_raw_poke(p_vec, loc, elt);
+}
+static inline
+void r_dyn_chr_push_back(struct r_dyn_array* p_vec, r_obj* elt) {
+  KEEP(elt);
+  r_ssize loc = r__dyn_increment(p_vec);
+  r_dyn_chr_poke(p_vec, loc, elt);
+  FREE(1);
 }
 static inline
 void r_dyn_list_push_back(struct r_dyn_array* p_vec, r_obj* elt) {