+++ src/rlang/env-binding.h
@@ -1,6 +1,9 @@
+// IWYU pragma: private; include "rlang.h"
+
 #ifndef RLANG_ENV_BINDING_H
 #define RLANG_ENV_BINDING_H
 
+#include "rlang-types.h"
 
 enum r_env_binding_type {
   R_ENV_BINDING_TYPE_value = 0,