+++ src/rlang/vec-chr.h
@@ -1,7 +1,10 @@
+// IWYU pragma: private; include "rlang.h"
+
 #ifndef RLANG_VECTOR_CHR_H
 #define RLANG_VECTOR_CHR_H
 
-#include <string.h>
+#include "globals.h"
+#include "rlang-types.h"
 
 
 static inline