+++ NEWS.md
@@ -1,5 +1,18 @@
 # rlang (development version)
 
+* `env_browse()` and `env_is_browsed()` are now defunct as they require an API
+  that is no longer available to packages (#1727).
+
+* `env_unlock()` is now defunct because recent versions of R no long
+  make it possible to unlock an environment (#1705). Make sure to use an
+  up-to-date version of pkgload (>= 1.4.0) following this change.
+
+
+# rlang 1.1.4
+
+* Added missing C level `r_dyn_raw_push_back()` and `r_dyn_chr_push_back()`
+  utilities (#1699).
+
 * `last_trace()` hyperlinks now use the modern `x-r-run` format (#1678).
 
 
@@ -1217,7 +1230,7 @@ extensive changes to the display of error messages.
 
 * Infix operators now stick to their LHS when deparsed by
   `expr_deparse()` (#890).
-  
+
 
 # rlang 0.4.2
 
@@ -2002,7 +2015,7 @@ error reporting, tidy eval, and tidy dots.
 * `env_get()` now evaluates promises and active bindings since these are
   internal objects which should not be exposed at the R level (#554)
 
-* `env_print()` calls `get_env()` on its argument, making it easier to 
+* `env_print()` calls `get_env()` on its argument, making it easier to
   see the environment of closures and quosures (#567).
 
 * `env_get()` now supports retrieving missing arguments when `inherit`